Validate outputFormat query parameters
This commit allows outputFormat to only be one of a known set of format types.  Formats are defined by the directories of Twig templates that are available.

The outputFormat string had been used as a part of the file path for Twig templates.  This meant people could inject strings into the filepath an look around on the hard drive of the server.